Test getData, test not passing
