+++ package.json
@@ -31,6 +31,7 @@
     "@babel/register": "^7.9.0",
     "assertion-error": "^1.0.1",
     "chai": "^3.4.0",
+    "chai-http": "^4.3.0",
     "deep-eql": "^0.1.3",
     "jsdoc": "^3.6.6",
     "mocha": "^7.1.1",
